Add setupAllDataSources command
