+++ sema4ai/vscode-client/src/robocorpCommands.ts
@@ -150,3 +150,4 @@ export const SEMA4AI_DROP_DATA_SOURCE = "sema4ai.dropDataSource";  // Remove Dat
 export const SEMA4AI_SETUP_DATA_SOURCE = "sema4ai.setupDataSource";  // Setup Data Source
 export const SEMA4AI_OPEN_DATA_SOURCE_DEFINITION = "sema4ai.openDataSourceDefinition";  // Open Data Source definition
 export const SEMA4AI_DROP_ALL_DATA_SOURCES = "sema4ai.dropAllDataSources";  // Remove All Data Sources
+export const SEMA4AI_SETUP_ALL_DATA_SOURCES = "sema4ai.setupAllDataSources";  // Setup All Data Sources
